переписываем реген энергий у петов и мобов на реген энергии у мобов и петов теперь влияет хаста в статы петам добавляем возможность передавать модификатор урона от урона чара Реализуем поддержку новой системы Real Proc Per Minute (RPPM) раньше был берд от БК наверное Вспомогательные спелы с атрибутом SPELL_ATTR6_NO_DONE_PCT_DAMAGE_MODS так же не должны критовать Переделываем флаги при ауре смены фракции(флаги по офу выставил, но клиент ен понимает хз почему(((), но на пве теперь донимать людей не будут корректировки по по спелам друля реализуем работу Книги веков на острове

+++ src/server/game/AI/CreatureAI.h
@@ -148,7 +148,7 @@ class CreatureAI : public UnitAI
         virtual void OwnerAttacked(Unit* /*target*/) {}
 
         // Called when a creature regen one of his power
-        virtual void RegeneratePower(Powers power, int32& value) {}
+        virtual void RegeneratePower(Powers power, float& value) {}
 
         virtual void RecalcStats() {}
